comments.py: stop the 'same IP address' verifying on comment submission
DWiki signs the comment form with the /24 of the IP address it was
fetched from and some other stuff, and refuses to accept a comment
posting attempt if it comes from an address outside the /24. Well.
Used to, because this change takes that verification out; it was
not catching spam and it does seem to have caught real people trying
to post comments on Wandering Thoughts.