Allow directories to gently disallow being displayed in certain views
For a not so hypothetical example, you could have a blog directory with
a lot of entries that you'd rather avoid being displayed in a blogdir
view. Since this is gentle, visitors get a redirection to the default
view instead of a permission denied message.

This is first pass code. Things that may change in the future:

- disallowed views are not inherited by subdirectories
- Atom feeds and other non-public directory views can be disallowed
- trying to disallow the default view of a directory results in
  the default view being omitted from the page tools 'See As' links
  if you are viewing the directory in a non-default view.