DWikiText rendering: add (optional) word-wrapped preformatted text
I've added a new processing note that lets preformatted text be
word-wrapped (instead of unwrapped, the normal state of <pre>s).
Word-wrapped preformatted text avoids scrollbars when it's wide
and is often more readable, and it turns out that it's almost
always what I want for preformatted text today.

(I have spent quite a lot of time faking it badly.)

This approach is a wimpy half measure. Word-wrapping preformatted
text really should be the default, but that's a formatting change
(but I probably should still make it anyways).