context.py: add disk caching of 'find all descendants'
context.cache_page_children() now stores its result (for real
directories) in the on-disk cache as a heuristically-invalidated cache
entry. This is a bit sleazy (especially because invalidation is a bit
tricky) but is very much worthwhile for WanderingThoughts.

Today's moral: filesystem walks do not scale up really well. What
was fast with a hundred files is not necessarily so much with a few
thousand.

- atomgen.py changed to use context.cache_page_children() to take
  advantage of this cache
- documented in test/pages/dwiki/Caching

This is not the full change but is the easily-separated bit. Bad me.