Add a TitleIndex wikitext macro that's basically blog:titleindex
I've always wanted wikitext to be able to do an attractive reverse
chronological listing of pages (with dates and page titles, as you get
for date based virtual directories that are too large). So now I've
added it as the TitleIndex macro. In fact it literally just wraps
blog::titleindex.

In the process we turned out to need some CSS changes, because
wikitext CSS used to identify 'wiki tables' as 'any <td> inside
the wikitext <div>' and that now applies to the blogtitles table
produced by blog::titleindex, which has (or should have) completely
separate CSS styles. So now wikitext tables have an explicit class
and the CSS uses that class.

Also changed: RecentChanges and AllPages now use the disk cache
of page descendants. Also added a comment about how AllPages is
actually a terrible macro after all if you want a subset of pages.