macros.py: Add AllPagesUnder macro plus some documentation changes
As kind of threatened in the last commit, I've added an AllPagesUnder
macro. It turns out that not only is it more efficient but I can also
make it look nicer.

While I was looking at things, various bits of macro documentation
got touchups, DocAll and EnumerateAll got the ability to deal with
just text macros (so one can always get a definitive list of them),
and DocAll now marks the resulting page as non-cacheable (because
it's annoying to have your new macro documentation not show up due
to the old page being cached in the renderer cache).