Provide a simple way to configure an external dwiki.css
It's better to serve dwiki.css outside of DWiki itself, but the achilles
heel of doing so has always been that it required changing the standard
templates/html/css template. Fortunately there is an obvious way around
this (that only occurred to me recently, silly me); one can now set
'cssurlprefix' in your wiki.conf, and if set it's taken as the URL prefix
for dwiki.css. If unset, html/css keeps assuming ${staticurl}/dwiki.css
as it used to.