dwconfig.py: allow people to skip supplying an options object for materialize()
If you are trying to run DWiki as a real WSGI application under a WSGI
server, materialize() is the main entry point for creating the ready to
go application object. It turns out that it's convenient to not have to
make up an options object when you do this, so now we allow this.

(Easily running DWiki as a WSGI app requires a writeup and some more
cleanups, but baby steps.)