templates/dwiki/overrides.tmpl: allow for page-type-specific overrides
The short form version of why is that it's useful to be able to override
the normal view rendering for just files, not directories. Before
dwiki/overrides.tmpl changed, this required overriding the full
rendering for the normal view just to get access to the file-based one.
This is kind of silly when we already have a perfectly good set of
mechanisms (used in other places in the templates) for type-based stuff.