Significantly reform what wikitext rendering generates (first pass)
NOTE: this involves an incompatible change to the {{CutShort}} macro.
See later.
NOTE 2: this changes what's stored in the on disk render(er) cache.
Much of your old/current renderer cache will no longer be used after
this; you may want to delete it entirely and let things restart from
scratch.

Originally, rendering wikitext into HTML just produced HTML. Over
time it became obvious that this was a design mistake; rendering
wikitext also produces other information (like the title, if any)
and wedging all of these into various special cases was increasingly
awkward. Now I've changed that.

Rendering wikitext now produces a RendResult object that encapsulates
the result of the rendering. This RendResult (and an accompanying
titleInfo dictionary of title information) is what is cached on disk.
Actual template renderers get a RR object and turn it into HTML by
calling an appropriate function. Since the RR object contains a stream
of generated HTML blocks there are now a number of new renderers that
do things like render wikitext without the title or render the full
HTML version of the title, complete with <hN> element.

(Since variations of these renderers are sprouting like rabbits,
specific renderers are highly likely to change in the future. I
think I'm going to have to give in and create renderer arguments.)

In order to cache a single RR object that's useful across multiple
contexts, the implementation of {{CutShort}} had to change. Instead
of aborting the rendering when it triggers, {{CutShort}} now merely
adds a marker in the stream of generated HTML blocks; the full page
is always generated and always stored in the RR object. Turning the
RR object into HTML checks for CutShort markers when appropriate.
This creates two incompatible changes in {{CutShort}}, both of which
I think no one will notice:

* {{CutShort}} is now only valid at the beginning of a non-indented,
  non-nested line. This matches how I've always used it, but means
  that you can no longer cut a paragraph in the middle or the like.
  This is due to technical limitations of the implementation and may
  change.

* since the entire page is now always generated, access restrictions
  that were below a {{CutShort}} will now always trigger. It used to
  be that you could have a non-blocked teaser, a {{CutShort}}, and
  then an access restriction that would block full-page rendering.
  Not any more. Again, this was an obscure and undocumented feature
  (and probably not a good idea anyways).

This change removes wikirend.registerCached(), which required a change
to atomgen.py to stop using it. I've also added a new renderer,
atom::pageterse:notitle; this uses the new wikitext rendering stuff to
skip the title of a page when turning it into an Atom feed entry.