Template updates to take advantage of the new rendering
The big change is that 'blog entries' now show the entry date below the
title (if the entry has a title; if not the date is shown just above the
entry text). This makes entry dates more visible, which is something
that I've come to believe is important in a blog. Doing this requires
a CSS update, so there is a new version of test/statics/dwiki.css.

It also switches Atom feed generation to generating entry HTML that does
not include the title. Atom feed entries already include the title as an
Atom <title> entity, so this stops them from stuttering.