Add basic support for bypassing the IMC when you post a comment
In an environment where the in-memory cache is only used when DWiki
is under high load, it's sort of okay to make someone who posts a
comment wait for the IMC TTL to expire before their comment shows up.
In an environment where the IMC is in use all the time (because DWiki
is always running in a preforked environment), it's irrirating.
This commit adds somewhat basic support for getting around this
problem, plus some discussion of it in test/pages/dwiki/Caching.

- comments.py: set a temporary cookie when a comment is posted
  successfully. This implicitly bypasses the IMC and BFC, since
  neither of them intercept requests with cookies set (a feature
  that was originally intended to apply to logged-in users).

- httpcore.py: clear the IMC if a non-cacheable request generates
  a response with a cookie set. This is most frequently going to
  apply to POST requests and thus to posting a comment (where
  comments.py sets the cookie).
  This IMC clearing only applies to the current process, but in a
  low-load situation that process is hopefully the only process
  that's actually handling requests and so the only one with anything
  in the IMC.

(dwiki-scgi.py deliberately tries to reuse the same process in low load
situations, but other preforking environments may not behave the same.)