Add basic support for operating in a HTTPS environment
It turned out that I had hard-coded 'http:...' in various places, which
just doesn't work when DWiki was running in a web server environment
that was using HTTPS. Recognizing HTTPS is done by looking for the
apparent standard environment variable setting of 'HTTPS=on', although
this may not work in all web servers (it does in Apache and lighttpd
CGI). More research is probably called for here.

Future work: adding features so that DWiki can force https URLs in Atom
feeds and force http to https redirections if desired.