httpcore.py: detect and reject mangled Host: values
Mangled Host: values (taken either from Host: or from the request URI) can
be used to do various bad things.

This change also corrects the handling of hostnames in the request URI to
fully work once again. We now have to fix both reqdata['server-name'] and
the newly-added-for-https reqdata['server-url']. Probably this should be
restructured.