Support templates marking themselves as permission errors; template update
Added a renderer and magic support for it in views.py so that templates
can mark themselves as having permission errors; this results in a
proper HTTP status 403. Updated needauth.tmpl and comment/needauth.tmpl
to use this.

Updated errors/*.tmpl to all use ${:error:code} instead of hard-coding
404s, since they don't always generate 404s any more.

(Yes, multiple changes in one, I know. It's because of test.timestamps.)