Fix wikitext caching; make BFC and IMC more correct for http vs https
I accidentally broke most wikitext caching by having things call
_render() instead of _render_cached().

The code can now function properly in the face of a site available
over both http and https. This requires two things: the BFC and
IMC must use the URL schema as well as the host (we only include
it if it's https, http is the default), and absolute-links 'terse'
wikitext rendering (used for Atom feeds) must be stored in a key
that includes the schema (again http is the default and https is
marked).

This is somewhat of a hack. A better version would try to fully
consider which bits of the renderer cache have host-specific
things in them and which bits don't.