New configuration option 'canon-host-url'
This is the canonical URL (including schema and if necessary port)
for the DWiki's host. If set, it's used to generate redirections and
especially absolute URLs during wikitext rendering.

Suppose that you are doing a HTTP to HTTPS redirection and for some
reason you have feed fetchers that do not support that (hi FeedBurner).
In order to not break the feeds you must keep serving them via HTTP.
However you want the absolute URLs for entries and so on in the feed
to be the HTTPS URLs, not the HTTP URLs. This is what canon-host-url
is there for.

(As a side effect you will no longer have a vestigial and useless
DWiki cache for your HTTP-side DWiki.)

It goes without saying that this is kind of a hack. In an ideal
world it wouldn't be necessary because everyone would respect
your redirections and support HTTPS.