Fix templates/struct/breadcrumbs to allow an easy prefix override
In many DWiki setups the DWiki URL space is not at the root of the
overall URL space and you want some constant prefix on the breadcrumbs
to reflect that. This previously required changing the standard
breadcrumbs template itself. Now the template itself does the obvious
thing of supporting an Overrides/breadcrumbs-prefix template that is
inserted if it exists. (No such template is supplied by default.)