Add support for RSS 2.0 format feeds for recently changed pages
This is an unpleasant hack but it is less unpleasant to add direct
support for RSS 2.0 than to depend on external feed conversion services.
This support is limited in various ways, per NewFeatures/RSS2Feeds.
It is basically just good enough to give feed URLs to people who insist
on RSS 2.0 only (yes, there are some such people).