httpcore: add rss2 as another feed view
I've now remembered to update a few places that know about atom and
atomcomments to also include the new 'rss2' view. Probably this means
that there should be a single feed_views variable but that's for later
when I've thought about the issue more; this is a quick fix for now.