Atom comments: get rid of the By in 'By <name|login>' in author info
It looks better in the Atom <name> tag if we just give people's names
instead of 'By <...>', especially now that we have proper names and so
on.

All of this template stuff is unfortunately getting messier and messier
(ie, we are getting more and more templates). I may need conditionals in
templates just to cut down the template proliferation.