Revise show-comment template to not nest <a> tags
The show-a-comment template has an <a> anchor tag that used to wrap
the entire author information. Well, the author information can now
include a link of its own, so that wrapping was no longer valid. The
revised templates shift it to wrap just the date information (which
doesn't include a link, at least currently).