comments.py: change what .also lines in the password file do
They now control what is shown as name/etc for known users if there is
nothing recorded for these in the comment itself, rather than serving
as default values when comments are composed. As such they can be
retroactive. This also involves a rendering style change if there's a
login and a default name to more clearly distinguish them from normal
names.

(This is badly described.)