Allow known user URLs to be DWiki pages (and more)
Authenticated users can have a 'URL' specified for them in the DWiki
passwd file. Originally this had to be a full URL, but that's limiting.
Now we support anything that can normally be put into a [[...]]
DWikiText link.

This also changes the rendering of authenticated users a bit and
moves some wikirend code out from the great big rendering class.