pageranges.py: don't match clearly bad months or days
Previously, a request for a DWiki url like '.../2012/99/' would be
considered as a year+month request and then blow up with an exception
inside the calendar module (because '99' is in no way a valid month).
Now we exclude clearly invalid months or days so that URLs with them
are not considered to be dates/date ranges at all.