Switch code from using the old 'raise E,MSG' to the modern 'raise E(MSG)'
Partly because DWiki is quite old, its code started out using the old
Python exception raising syntax of 'raise CLASS, STR-MSG' and was never
updated to 'raise CLASS(STR-MSG)'. This commit fixes that, hopefully
without any errors.

(Regexp based changes can misfire and this is all in error paths that
are hard to test and trigger rarely, if at all.)