Explicitly open a few binary files in binary mode
DWiki has a few files that contain binary contents instead of text,
primarily the global-authseed file and anything in any of its caches
(which are all pickled data). I used to just open them in text mode
because it didn't make a difference. Now the code explicitly opens them
in binary mode because this better prepares DWiki for a hypothetical
future port to Python 3.

(If DWiki works on Windows I'll be amazed and it's certainly not a
coding goal, so I don't consider that a reason to use binary mode on
opens.)