wikirend.py: add a comment about my future desires with numbered lists
I would like to get rid of or narrow the '<digit> ...' format of
numbered lists, but I do not have a good solution for now. After
various flailing and failed experiments, it's time for a comment
for the future. (Two comments, since there are two places that
would need to change.)