wikirend.py: allow '> ' quotes in lists
We did not previously allow:

	* a list entry with:
	> a quote in the list

Now we do. The comment here suggests strongly that Daniel Martin saw
the potential desire for this years ago but faithfully followed the
limitations of my original wikitext renderer. Well, it's long since
time I fixed that, so now I have.

(At this point I need to rewrite the wikitext documentation to
conform to current reality, but for now I am committing the code
change alone.)