wikirend.py: hack around the accidental numbered lists in paragraph trap
This trap is that if you write:

	A paragraph with a wrapped line talking about version
	1 of something ...

The '1' is (or was) interpreted as the start of a numbered list, not
just another line of the paragraph. Our workaround is an extreme hack;
when ingesting paragraph lines, if we hit what the initial tokenizer
sees as an 'ol' we check to see if the actual character is a '#'. If it
isn't, we take the 'ol' as a plain line of text instead.

This is a backwards incompatible change that required some content
rewrites in Wandering Thoughts, but I don't care; this case is annoying
enough when I hit it that I want it gone and I'm willing to hack things
up.

(At least this is my current story. I reserve the right to change my
mind later.)