wikirend.py: add a hack to not code-ify 'an_identifier_like_this'
Although '_' remains generally matching, we no longer accept it as
starting a <code> block if it immediately follows a letter or a digit.
Our definition of 'letter or digit' is naive in a Unicode world, but
that's life.

This is a hack, and perhaps it should also require that there be a letter
or digit following the _ as well. However, even as a hack it fixes an
extremely common thing that people writing DWikiText content do.

Unfortunately it's impossible to fix rendering so that '_a_identifier_'
works, at least with my current approach. Ultimately this is because I
have actually written '_struct_s' in content and expected that to come
out like it looks.