storage.py: improve write handling for durability and errors
In an attempt to make newly written things stick (such as comments),
we now fsync() everything in sight on write, both the file itself and
the containing directory (because that's sometimes been necessary).

In addition, we now explicitly close the file if there was a write
error, or more precisely if there was an exception raised in the write
phase (which now covers the fsync as well). This adds a final 'close'
phase as a necessary consequence.