httpcore.py: add yet more normally seen URL query parameters
There are a steadily increasing profusion of random query parameters
that people staple on URLs these days. This expands the list of them
that we attempt to deal with. Probably I should just give up on this
and at least add a configuration option that accepts any garbage
parameter, at least on GETs.