comments.py: update the zen block-on-result list of IPs
The Spamhaus ZEN DNSBL returns various different IPs depending on what
sort of badness the IP you queried had, and we don't want to block on
all of them (to the extent that I want to block on Zen listings at
all, which I currently don't). At some point I updated the list of
result IPs we use and wrote comments for them, and that's what this
is.