comments.py: update the collection of banned comment contents
Mostly this is an assortment of URLs, and I should remove most of them
because they're probably obsolete now (as a practical matter they were
probably mostly obsolete when added, as I think that most spammers
don't come back). The important change is to block '<a href="https://'
as well as the http:// version, since more and more things are moving
to HTTPS now.