Fix struct/head.tmpl template to work when there is no page
Under some circumstances, such as doing a search, there is no current
page and the '${page}' expansion in this template piece was failing.
I'm not sure how I managed to overlook this in the master source for
so long, but it was fixed long ago in the live DWiki instances I run.
This is the minimum touch-up of the template to make search work
again, pending a larger update to more closely match current reality.