macros.py: update IMG's help text to cover using 'auto' for height
The modern usage of the IMG macro specifies the height as 'auto', so
that we generate an inline style to auto-scale things. While I'm here,
also mention which of the alt text and the title text is shown by
browsers when you hover over the image (it's the title text).