macros.py: add listing pages with matching words in their page name
This adds a new macro, MatchingPagesUnder (and a test file for it)
that allows listing pages under a particular directory that have a
matching word (or sequence of words) in their page name. This is a
low-rent way of implementing tags/topics, if you either have a
relevant word in your page names or can just list the page names. We
also support an exclusion list, so that in the case where your initial
matching word is a common one (like 'Go'), you can exclude some pages
that aren't actually on your topic.

The implementation is a bit of a hack and it is currently slightly
dangerous in that it allows regular expression characters.